The software now has two processing platforms: CPU and GPU. I've successfully completed the first HarmonicBondForce Cuda implementation for a large system of 23k JAC protein atoms. Results show a computational time that is nearly 10 times faster than the CPU implementation.

+++ core/baseline/headers/Reporter.h
@@ -60,7 +60,7 @@ namespace BaseLine {
         // Main function to read, process, and write PDB file
         //void pdbOutputGenerator(const std::string& inputFilename, const std::string& outputFilename, std::vector<PDBAtom>& outputTemplate,const std::vector<Coords3D>& positions, int step);
         void pdbOutputGeneratorPart1(const std::string& inputFilename, const std::string& outputFilename, std::vector<PDBAtom>& outputTemplate);
-        void pdbOutputGeneratorPart2(const std::string& outputFilename, std::vector<PDBAtom>& outputTemplate, const std::vector<Coords3D>& positions, size_t ModelNum);
+        void pdbOutputGeneratorPart2(const std::string& outputFilename, std::vector<PDBAtom>& outputTemplate, const std::vector<Coords3D>& positions, int ModelNum);
 
 
     private:
