+++ core/baseline/src/Initializer.cpp
@@ -29,7 +29,7 @@ using namespace BaseLine;
 //    return pair(totalForces, velocities);
 //}
 void Initializer::InitializeForcesAndVelocities(const vector<Coords3D>& atomPositions, vector<Coords3D>& totalForces, vector<Coords3D>& velocities) {
-    size_t numAtoms = atomPositions.size();
+    int numAtoms = atomPositions.size();
     totalForces.assign(numAtoms, Coords3D(0, 0, 0));
     velocities.assign(numAtoms, Coords3D(0, 0, 0));
